perfmon: fix perf_user_access_enabled type
perf_user_access_enabled is defined as u8,
clib_sysfs_read format type is %u, this is for unsigned int,
change type from u8 to u32.

Type: fix
Fixes: 268d7be66b8b ("perfmon: enable perfmon plugin for Arm")

+++ src/plugins/perfmon/arm/events.c
@@ -75,7 +75,7 @@ arm_init (vlib_main_t *vm, perfmon_source_t *src)
     - if the file exists and is 0, user access needs to be granted
       with 'sudo sysctl kernel/perf_user_access=1'
   */
-  u8 perf_user_access_enabled;
+  u32 perf_user_access_enabled;
   char *path = "/proc/sys/kernel/perf_user_access";
   err = clib_sysfs_read (path, "%u", &perf_user_access_enabled);
   if (err)
